refactor: organizes pages folders
